[HUBS] Fix "arguments" allocations in WebGLState
WebGLState wraps texImage2D, compressedTexImage2D, and texImage3D to
swallow exceptions. The upstream implementation uses "arguments" to
handle the various function signatures. This can allocate in some
javascript engines. Instead we hardcode the handful of signature
patterns based on missing arguments. Note that we only need to account
for the signatures with different arity, not type since things are
directly passed through.

This seems generally useful and we should try to land this upstream.