[HUBS] Reduce matrix update cost using dirty flags
This is a significant change to the way Three handles updating local and
world matrices for objects. It is not backwards compatible with
existing ThreeJS code for the most part and requires significant effort
on the part of the developer to make sure the contract is
followed (namely setting matrixNeedsUpdate and decomposing to
position/rotation/scale if the local matrix is manually updated.)

In exchange matrices will only be updated when absolutely necessary
which is a significant performance improvment, especially with deep
scene hierarchy which aframe exacerbates with its additional Group
nodes for every entity. In addition to dirty checking, matrixWorld is
shared with parent node when local matrix has not been modified form
identity, which allows skipping some matrix multiplications when
computing world matrices.

With this patch you must make sure to follow these rules or very strange things will happen.
- If you modify an object's position, rotation, quaternion, or scale you MUST set matrixNeedsUpdate.
- If you modify an object's matrix
      you MUST decompose() back onto its position, quaternion and scale and
      you MUST set matrixWorldNeedsUpdate (or matrixNeedsUpdate, but the former is more correct).
      (applyMatrix() and updateMatrix() handle this for you)
- If you modify an object's matrixWorld
      you MUST make sure it has previously been modified so that it is not using its parent matrix as its own,
      you MUST update its local matrix,
      you MUST update its position/quaternion/scale, and
      you MUST set childrenNeedMatrixWorldUpdate.
      (setMatrixWorld() handles all of this for you)
- Before you read an object's matrix you MUST call updateMatrix() or updateMatrices().
- Before you read an object's matrixWorld you MUST call updateMatrices().
      (getWorldPosition, getWorldOrientation and getWorldScale handle this for you)
- Do not set matrixIsModified yourself; You could accidentally overwrite a shared parent matrixWorld.
- Note updateMatrix, updateMatrixWorld, updateWorldMatrix, updateMatrices, setMatrixWorld,
      matrixNeedsUpdate, matrixWorldNeedsUpdate, and matrixIsModified are all different things.
      Most already exist in ThreeJS but some have been added here.
      Double check you are using the one you intend to.
      Be on the lookout for compatibility issues with third party libraries.