[HUBS] Fix a no matrices update bug (#65)
* Fix a invisible object bug in Object3D.updateMatrixWorld()

Object3D.updateMatrixWorld() skips the update of invisible
objects unless includeInvisible parameter is true.

Regardless of an object's visibility, its parent's
childrenNeedMatrixWorldUpdate is cleared if
Object3D.updateMatrixWorld() is recursively called from
the parent.

In that case the parent's childNeedMatrixWorldUpdate is
cleared and the object's matrixNeedsUpdate or
matrixWorldNeedsUpdate isn't set (unless it's manually set)
so that the object may lose the opportunity to update
the world matrix. As a result the object can be rendered at
a wrong place in the scene when it becomes visible.

This commit resolves the problem by raising the
matrixWorldNeedsUpdate flag instead in that case,
and the object's world matrix can be updated later when needed.

* Add matrixNeedsUpdate and visibility unit test