[HUBS] [Upcoming] FileLoader: HTTP Range requests support (#68)
[HUBS]
This change is needed for glTF LOD progressive loading.
This change is not merged to the official Three.js PR yet
but likely it can get in soon.

https://github.com/mrdoob/three.js/pull/24580

We can remove this commit if the PR is merged to
the official Three.js and we upgrade our Three.js fork to
the one including the change.