[HUBS] [Upcoming] GLTFLoader: Add loadNode hook
[HUBS]
This change is needed for glTF LOD support.
This change is not merged into the official Three.js but
likely it can get in soon.
https://github.com/mrdoob/three.js/pull/25077
We can remove this commit if the PR is merged to the
official Three.js and upgrading our Three.js fork to the
one including the change.